Refactor RoomOccupancyCalculation
change field type from int -> long

+++ src/main/java/com/zonkil/roomoccupancymanager/service/DefaultRoomOccupancyService.java
@@ -43,9 +43,9 @@ public class DefaultRoomOccupancyService implements RoomOccupancyService {
 				economyGuestCount - upgradeData.numberOfGuestToPromote);
 
 		return RoomOccupancyCalculation.builder()
-									   .premiumOccupancy((int) premiumOccupancy)
+									   .premiumOccupancy(premiumOccupancy)
 									   .premiumTotalMoney(premiumProfit)
-									   .economyOccupancy((int) economyOccupancy)
+									   .economyOccupancy(economyOccupancy)
 									   .economyTotalMoney(economyProfit)
 									   .build();
 	}
