+++ src/main/java/com/zonkil/roomoccupancymanager/web/dto/RoomOccupancyCalculationResponseDto.java
@@ -10,11 +10,11 @@ import java.math.BigDecimal;
 @Builder
 public class RoomOccupancyCalculationResponseDto {
 	@Schema(description = "Number of occupied premium rooms", example = "3")
-	private int premiumOccupancy;
+	private long premiumOccupancy;
 	@Schema(description = "Profit from premium rooms", example = "738")
 	private BigDecimal premiumTotalMoney;
 	@Schema(description = "Number of occupied economy rooms", example = "3")
-	private int economyOccupancy;
+	private long economyOccupancy;
 	@Schema(description = "Profit from economy rooms", example = "167.99")
 	private BigDecimal economyTotalMoney;
 }
